@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "main": "build/server.js",
+  "type": "module",
   "scripts": {
     "start": "npm run run:dev",
     "run:dev": "node config/build.js IS_LOCAL=true ENV=development",
@@ -29,9 +30,11 @@
     "react-dom": "^18.1.0",
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
-    "rxjs": "^7.5.5"
+    "rxjs": "^7.5.6"
   },
   "devDependencies": {
+    "@swc/core": "^1.2.244",
+    "@swc/jest": "^0.2.22",
     "@types/express": "^4.17.13",
     "@types/express-http-proxy": "^1.6.3",
     "@types/jest": "^28.1.4",
@@ -63,7 +66,7 @@
     "jest-environment-jsdom": "^28.1.2",
     "mddir": "^1.1.1",
     "mini-css-extract-plugin": "^2.6.1",
-    "node-sass": "^4.14.1",
+    "node-sass": "^6.0.1",
     "nodemon-webpack-plugin": "^4.7.1",
     "pm2": "^5.2.0",
     "prettier": "^2.7.1",
@@ -73,12 +76,13 @@
     "shelljs": "^0.8.5",
     "stream-browserify": "^3.0.0",
     "style-loader": "^3.3.1",
+    "swc-loader": "^0.2.3",
     "terser-webpack-plugin": "^5.3.3",
     "ts-jest": "^28.0.5",
-    "ts-loader": "^9.3.0",
+    "ts-jest-resolver": "^2.0.0",
     "ts-node": "^10.8.1",
     "typescript": "^4.7.3",
-    "webpack": "^5.73.0",
+    "webpack": "^5.74.0",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.9.2",
     "webpack-merge": "^5.8.0",
