@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { ContextData } from "core/models/context.model";
+import { ContextData } from "core/models/context.model.js";
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
