@@ -1,8 +1,8 @@
 import React from "react";
 import { Observable } from "rxjs";
-import { ContextData } from "src/core/models/context.model";
-import { IRedirect, PageData } from "src/core/models/page-data";
-import { ApiResponse } from "src/core/models/api-response";
+import { ContextData } from "src/core/models/context.model.js";
+import { IRedirect, PageData } from "src/core/models/page-data.js";
+import { ApiResponse } from "src/core/models/api-response.js";
 
 declare global {
     interface Window {
